GO/scripts/data-migrator/requirements.txt
cassandra-driver==2.7.2
pymongo==2.7.2